djblog/djblog/requirements.txt
asgiref==3.5.0
Django==4.0.3
sqlparse==0.4.2
tzdata==2022.1
django-phonenumber-field[phonenumbers]
djangorestframework
Pillow==9.0.1